backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic-settings>=2.5.0
python-dotenv>=1.0.0
aiofiles>=24.1.0
python-docx>=1.1.0
PyPDF2>=3.0.0
lxml>=5.0.0
llama-cpp-python>=0.3.0